# swagger.yaml
openapi: 3.0.0
info:
  title: gat-tools
  description: Ferramenta de automação para processos da Gestão de Ambientes de Testes.
  version: 0.0.1
servers:
- url: /api/v1
  description: Desenvolvimento local
paths:
  /datasources:
    get:
      tags:
      - ASM
      summary: Returns a list of datasources
      operationId: listDatasources
      parameters:
      - name: type
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - app
          - asm
      - name: applicationCode
        in: query
        required: false
        style: form
        schema:
          type: string
      security:
        - basicAuth: []
      responses:
        200:
          description: A JSON array of datasource
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/datasource'
                x-content-type: application/json
        401:
          $ref: '#/components/responses/UnauthorizedError'
      x-swagger-router-controller: DatasourceController
  /applications:
    get:
      tags:
      - Application
      summary: Returns a list of applications
      operationId: listApplications
      parameters: []
      responses:
        200:
          description: A JSON array of application
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/application'
                x-content-type: application/json
      x-swagger-router-controller: ApplicationController
  /netsms-version/{datasourceCode}:
    get:
      tags:
      - NETSMS
      summary: Returns a version Object
      operationId: listNetsmsVersion
      parameters:
      - name: datasourceCode
        in: path
        required: true
        explode: true
        schema:
          type: string
      responses:
        200:
          description: A JSON array of application
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/netsms-version'
                x-content-type: application/json
      x-swagger-router-controller: NetsmsVersionController
  /netsms-parameter/{datasourceCode}:
    get:
      tags:
      - NETSMS
      summary: Returns a version Object
      operationId: listNetsmsParameter
      parameters:
      - name: datasourceCode
        in: path
        required: true
        explode: true
        schema:
          type: string
      responses:
        200:
          description: A JSON array of application
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/netsms-parameter'
                x-content-type: application/json
      x-swagger-router-controller: NetsmsParameterController
  /cenarios:
    get:
      tags:
      - Application
      summary: Returns a list of cenarios
      operationId: listCenarios
      parameters: []
      responses:
        200:
          description: A JSON array of cenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/cenario'
                x-content-type: application/json
      x-swagger-router-controller: ApplicationController
  /environments:
    get:
      tags:
      - Application
      summary: Returns a list of environments
      operationId: listEnvironments
      parameters: []
      responses:
        200:
          description: A JSON array of environments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/environment'
                x-content-type: application/json
      x-swagger-router-controller: ApplicationController
  /accesscontrols/{datasourceCode}:
    get:
      tags:
      - AccessControl
      summary: Return a list of Access Controls.
      operationId: listAccesscontrols
      parameters:
      - name: datasourceCode
        in: path
        required: true
        explode: true
        schema:
          type: string
      - name: type
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - logon
          - monitora
      responses:
        200:
          description: A JSON array of Access Control objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/accesscontrol'
                x-content-type: application/json
      x-swagger-router-controller: AccesscontrolController
  /asm-disks/{datasourceCode}:
    get:
      tags:
      - ASM
      summary: Returns a list of ASM disks.
      operationId: listAsmdisks
      parameters:
      - name: datasourceCode
        in: path
        required: true
        explode: true
        schema:
          type: string
      responses:
        200:
          description: A JSON array of ASM Disk objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/asmdisk'
                x-content-type: application/json
      x-swagger-router-controller: AsmController
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    datasource:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
      example:
        type: asm
        code: net001ht4sun002b
    application:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        description: NetSMS (legado)
        code: NETSMS
    cenario:
      type: object
      properties:
        code:
          type: string
        applicationCode:
          type: string
        description:
          type: string
          nullable: true
      example:
        description: Teste Integrado - lado direito
        applicationCode: NETSMS
        code: SIT1
    environment:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      example:
        description: Teste Integrado - lado direito
        code: SIT4
    asmdisk:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        diskgroupNumber:
          type: integer
        diskgroupName:
          type: string
        instanceName:
          type: string
        databaseName:
          type: string
        softwareVersion:
          type: string
        hostName:
          type: string
      example:
        name: BRA_GOLDEN_102017_0000
        path: /dev/rdsk/c0t60A9800041765A706D5D4673544F6257d0s6
        databaseName: BRASAMX
        diskgroupNumber: 11
        instanceName: +ASM        
        softwareVersion: 11.2.0.4.0
        diskgroupName: BRA_BRASAMX_DATA102017
    accesscontrol:
      type: object
      properties:
        owner:
          type: string
        name:
          type: string
        valid:
          type: boolean
        enabled:
          type: boolean
        type:
          type: string
      example:
        owner: CORE
        name: TR_LOGON
        valid: true
        enabled: true
        type: logon
    netsms-version:
      type: object
      properties:
        version:
          type: string
        applyDate:
          type: string
          format: date-time
        executableId:
          type: integer
      example:
        version: 9.4.1
        applyDate: 2019-04-17T19:42:39.000Z
        executableId: 400
    netsms-parameter:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        numValue:
          type: number
      example:
        name: 9.4.1
        description: 2019-04-17T19:42:39.000Z
        numValue: 400